Updated File Structure & Finished Book Details Feature
Modified file structure to be entity based. This makes navigating through the project a lot easier in my opinion. Finished writing code for Book Details feature and wrote a postman test suite for it.
