+++ src/main/java/com/example/bookstoreProject/Author/Author.java
@@ -1,4 +1,4 @@
-package com.example.bookstoreProject.entity;
+package com.example.bookstoreProject.Author;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -21,20 +21,21 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "users")
-public class User {
-
+@Table(name = "authors")
+public class Author {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String user_name;
-    private String password;
     private String first_name;
     private String last_name;
-    private String email;
+    private String bio;
+    private String publisher;
     @CreationTimestamp
     private LocalDateTime date_created;
     @UpdateTimestamp
     private LocalDateTime date_updated;
-}
 
+    public Author(Long id) {
+        this.id = id;
+    }
+}
