+++ src/main/java/com/example/bookstoreProject/Author/AuthorRepository.java
@@ -1,8 +1,7 @@
-package com.example.bookstoreProject.repository;
+package com.example.bookstoreProject.Author;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.example.bookstoreProject.entity.Author;
 
 @Repository
 public interface AuthorRepository extends JpaRepository<Author, Long>{
