+++ src/main/java/com/example/bookstoreProject/Author/AuthorService.java
@@ -1,0 +1,5 @@
+package com.example.bookstoreProject.Author;
+
+public interface AuthorService {
+    Author createAuthor(Author Author);
+}
