+++ src/main/java/com/example/bookstoreProject/Book/BookRepository.java
@@ -1,10 +1,9 @@
-package com.example.bookstoreProject.repository;
+package com.example.bookstoreProject.Book;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.example.bookstoreProject.entity.Book;
 import org.springframework.data.jpa.repository.Query;
