+++ src/main/java/com/example/bookstoreProject/ShoppingCart/ShoppingCart.java
@@ -1,13 +1,17 @@
-package com.example.bookstoreProject.entity;
+package com.example.bookstoreProject.ShoppingCart;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
+
+import com.example.bookstoreProject.Book.Book;
+import com.example.bookstoreProject.User.User;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -21,21 +25,20 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "authors")
-public class Author {
+@Table(name = "shopping_cart")
+public class ShoppingCart {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String first_name;
-    private String last_name;
-    private String bio;
-    private String publisher;
+    @ManyToOne
+    User user;
+    @ManyToOne
+    Book book;
+    
+    
     @CreationTimestamp
     private LocalDateTime date_created;
     @UpdateTimestamp
     private LocalDateTime date_updated;
-
-    public Author(Long id) {
-        this.id = id;
-    }
 }
