+++ src/main/java/com/example/bookstoreProject/User/User.java
@@ -1,10 +1,9 @@
-package com.example.bookstoreProject.entity;
+package com.example.bookstoreProject.User;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 import org.hibernate.annotations.CreationTimestamp;
@@ -22,21 +21,20 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "credit_cards")
-public class CreditCard {
+@Table(name = "users")
+public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @ManyToOne
-    User user;
-    private int card_number;
-    private int zip_code;
-    private int cvv;
-    
-    
+    private String user_name;
+    private String password;
+    private String first_name;
+    private String last_name;
+    private String email;
     @CreationTimestamp
     private LocalDateTime date_created;
     @UpdateTimestamp
     private LocalDateTime date_updated;
 }
+
