+++ src/main/java/com/example/bookstoreProject/User/UserRepository.java
@@ -1,6 +1,4 @@
-package com.example.bookstoreProject.repository;
-import com.example.bookstoreProject.entity.User;
-
+package com.example.bookstoreProject.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.Query;
