--- src/main/java/com/example/bookstoreProject/repository/CommentRepository.java
@@ -1,15 +1,0 @@
-package com.example.bookstoreProject.repository;
-import com.example.bookstoreProject.entity.Comment;
-import com.example.bookstoreProject.entity.User;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    List<Comment> findByPost(Post post);
-
-    List<Comment> findAllByUser(User user);
-}
