--- src/main/java/com/example/bookstoreProject/service/AuthorService.java
@@ -1,7 +1,0 @@
-package com.example.bookstoreProject.service;
-
-import com.example.bookstoreProject.entity.Author;
-
-public interface AuthorService {
-    Author createAuthor(Author Author);
-}
