+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/bookstore
 spring.datasource.username=root
-spring.datasource.password=Jacorey561
+spring.datasource.password=Grubsitch!1
 
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL8Dialect
